# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Intania Shop API
  version: 1.0.0
  description: |
    Backend API for the ESC Chula Intania Shop.
    Identity comes from a JWT. Operations described as Admin only require the ADMIN role.
servers:
  - url: /
    description: Same host as this documentation
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [Products]
      summary: List products
      parameters:
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PageSize" }
        - name: include_variants
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200": { $ref: "#/components/responses/ProductList" }
        "500": { $ref: "#/components/responses/Error" }
    post:
      operationId: createProduct
      tags: [Products]
      summary: Create product
      description: Admin only.
      security: [{ bearerAuth: [] }]
      requestBody: { $ref: "#/components/requestBodies/CreateProduct" }
      responses:
        "201": { $ref: "#/components/responses/ProductDetail" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
...
  /products/{id}:
    parameters: [{ $ref: "#/components/parameters/ID" }]
    get:
      operationId: getProduct
      tags: [Products]
      summary: Get product details
      responses:
        "200": { $ref: "#/components/responses/ProductDetail" }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    put:
      operationId: updateProduct
      tags: [Products]
      summary: Update product
      description: Admin only. Supplied fields replace existing values.
      security: [{ bearerAuth: [] }]
      requestBody: { $ref: "#/components/requestBodies/UpdateProduct" }
      responses:
        "200": { $ref: "#/components/responses/ProductDetail" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteProduct
      tags: [Products]
      summary: Delete product
      description: Admin only.
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Product deleted }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /products/{id}/variants:
    parameters: [{ $ref: "#/components/parameters/ID" }]
    post:
      operationId: createProductVariant
      tags: [Products]
      summary: Create product variant
      description: Admin only.
      security: [{ bearerAuth: [] }]
      requestBody: { $ref: "#/components/requestBodies/Variant" }
      responses:
        "201": { $ref: "#/components/responses/Variant" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /variants/{id}:
    parameters: [{ $ref: "#/components/parameters/ID" }]
    put:
      operationId: updateVariant
      tags: [Products]
      summary: Update product variant
      description: Admin only.
      security: [{ bearerAuth: [] }]
      requestBody: { $ref: "#/components/requestBodies/Variant" }
      responses:
        "200": { $ref: "#/components/responses/Variant" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteVariant
      tags: [Products]
      summary: Delete product variant
      description: Admin only.
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Variant deleted }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
    Page:
      name: page
      in: query
      schema: { type: integer, format: int32, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, format: int32, minimum: 1, default: 10 }
  requestBodies:
    CreateProduct:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/CreateProductRequest" }
    UpdateProduct:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ProductInput" }
    Variant:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/VariantInput" }
...
  responses:
    Error:
      description: Request failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    ProductList: { description: Paginated products, content: { application/json: { schema: { $ref: "#/components/schemas/ProductListEnvelope" } } } }
...
    ProductDetail: { description: Product details, content: { application/json: { schema: { $ref: "#/components/schemas/ProductDetailEnvelope" } } } }
    Variant: { description: Product variant, content: { application/json: { schema: { $ref: "#/components/schemas/VariantEnvelope" } } } }
...
  schemas:
    SuccessBase:
      type: object
      required: [success]
      properties:
        success: { type: boolean, const: true }
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, const: false }
        error: { type: string, example: Product not found }
...
    ProductStatus: { type: string, enum: [Preorder, InStock, OutOfStock] }
    ProductType: { type: string, enum: [Single, Multiple] }
    PickupMethods:
      type: object
      properties:
        self_pickup: { type: boolean }
        home_delivery: { type: boolean }
    ProductListItem:
      type: object
      required: [product_id, name, base_price, status, images]
      properties:
        product_id: { type: integer, format: int64 }
        name: { type: string }
        base_price: { type: string, example: "299.00" }
        status: { $ref: "#/components/schemas/ProductStatus" }
        category: { type: [string, "null"] }
        images: { type: array, items: { type: [string, "null"], format: uri } }
        min_order: { type: [integer, "null"], format: int32 }
        max_order: { type: [integer, "null"], format: int32 }
    Variant:
      type: object
      required: [variant_id, product_id]
      properties:
        variant_id: { type: integer, format: int64 }
        product_id: { type: integer, format: int64 }
        size: { type: [string, "null"] }
        color: { type: [string, "null"] }
        stock_quantity: { type: [integer, "null"], format: int32 }
        price: { type: [string, "null"], example: "329.00" }
    ProductDetail:
      allOf:
        - { $ref: "#/components/schemas/ProductListItem" }
        - type: object
          required: [variants]
          properties:
            description: { type: [string, "null"] }
            stock_quantity: { type: [integer, "null"], format: int32 }
            preview_video: { type: [string, "null"], format: uri }
            shipping: { type: array, items: { type: [string, "null"] } }
            product_type: { oneOf: [{ $ref: "#/components/schemas/ProductType" }, { type: "null" }] }
            size_chart: { type: [string, "null"], format: uri }
            pickup_methods: { oneOf: [{ $ref: "#/components/schemas/PickupMethods" }, { type: "null" }] }
            pickup_location: { type: [string, "null"] }
            shipping_fee: { type: [string, "null"] }
            sku: { type: [string, "null"] }
            product_code: { type: [string, "null"] }
            created_at: { type: [string, "null"], format: date-time }
            updated_at: { type: [string, "null"], format: date-time }
            variants: { type: array, items: { $ref: "#/components/schemas/Variant" } }
    ProductInput:
      type: object
      properties:
        name: { type: string }
        description: { type: [string, "null"] }
        base_price: { type: string, example: "299.00" }
        status: { $ref: "#/components/schemas/ProductStatus" }
        category: { type: [string, "null"] }
        stock_quantity: { type: [integer, "null"], format: int32 }
        images: { type: array, items: { type: [string, "null"], format: uri } }
        preview_video: { type: [string, "null"], format: uri }
        shipping: { type: array, items: { type: [string, "null"] } }
        product_type: { $ref: "#/components/schemas/ProductType" }
        min_order: { type: [integer, "null"], format: int32 }
        max_order: { type: [integer, "null"], format: int32 }
        size_chart: { type: [string, "null"], format: uri }
        pickup_methods: { $ref: "#/components/schemas/PickupMethods" }
        pickup_location: { type: [string, "null"] }
        shipping_fee: { type: [string, "null"] }
        sku: { type: [string, "null"] }
        product_code: { type: [string, "null"] }
        variants: { type: array, items: { $ref: "#/components/schemas/VariantInput" } }
    CreateProductRequest:
      allOf:
        - { $ref: "#/components/schemas/ProductInput" }
        - { type: object, required: [name, base_price] }
    VariantInput:
      type: object
      properties:
        size: { type: [string, "null"] }
        color: { type: [string, "null"] }
        stock_quantity: { type: [integer, "null"], format: int32 }
        price: { type: [string, "null"] }
    ProductListData:
      type: object
      required: [products, total, page, page_size, total_pages]
      properties:
        products: { type: array, items: { oneOf: [{ $ref: "#/components/schemas/ProductListItem" }, { $ref: "#/components/schemas/ProductDetail" }] } }
        total: { type: integer, format: int64 }
        page: { type: integer, format: int32 }
        page_size: { type: integer, format: int32 }
        total_pages: { type: integer, format: int32 }
...
    ProductListEnvelope:
      allOf:
        - { $ref: "#/components/schemas/SuccessBase" }
        - { type: object, required: [data], properties: { data: { $ref: "#/components/schemas/ProductListData" } } }
...
    ProductDetailEnvelope:
      allOf:
        - { $ref: "#/components/schemas/SuccessBase" }
        - { type: object, required: [data], properties: { data: { $ref: "#/components/schemas/ProductDetail" } } }
    VariantEnvelope:
      allOf:
        - { $ref: "#/components/schemas/SuccessBase" }
        - { type: object, required: [data], properties: { data: { $ref: "#/components/schemas/Variant" } } }
...